feat(daemon): fjall storage backend for task/cron state (#2285)
Wave 4: adds fjall-backed task/cron storage behind default `fjall` feature.
SQLite preserved behind `sqlite` feature. Public API unchanged.

- Split `state.rs` into `state/fjall_store.rs` (default) and
  `state/sqlite_store.rs` (`sqlite` feature) using the same cfg-gate
  pattern as graphe (Wave 1).
- `TaskState` gains `Serialize`/`Deserialize` derives required for
  JSON encoding in the fjall backend.
- Added `Storage` and `StoredJson` error variants to `error.rs` for
  fjall-specific failure paths.
- Key schema: partition `ops:tasks`, key = `{task_id}`, value =
  JSON `TaskState`.
- Log event `task state restored FROM SQLite` updated to backend-agnostic
  `task state restored from store`.
- All 323 unit tests pass; workspace compiles clean.

Refs #2285

Gate-Passed: kanon 0.1.0
