+++ crates/daemon/src/error.rs
@@ -54,6 +54,24 @@ pub enum Error {
         location: snafu::Location,
     },
 
+    /// Storage backend error (fjall task-state store).
+    #[cfg(feature = "fjall")]
+    #[snafu(display("task-state storage error: {message}"))]
+    Storage {
+        message: String,
+        #[snafu(implicit)]
+        location: snafu::Location,
+    },
+
+    /// JSON serialization/deserialization error within stored task state.
+    #[cfg(feature = "fjall")]
+    #[snafu(display("task-state JSON error: {source}"))]
+    StoredJson {
+        source: serde_json::Error,
+        #[snafu(implicit)]
+        location: snafu::Location,
+    },
+
     /// I/O error during maintenance operation.
     #[snafu(display("maintenance I/O error: {context}"))]
     MaintenanceIo {
